+++ src/EverTask.Abstractions/IEverTaskHandlerOptions.cs
@@ -3,14 +3,20 @@ namespace EverTask.Abstractions;
 public interface IEverTaskHandlerOptions
 {
     /// <summary>
-    /// Gets or sets the retry policy for this task. Default is a linear retry policy with 3 tries every 500 milliseconds
+    /// Gets the retry policy for this task. Default is a linear retry policy with 3 tries every 500 milliseconds
     /// </summary>
-    public IRetryPolicy? RetryPolicy { get; set; }
+    public IRetryPolicy? RetryPolicy { get; }
 
     /// <summary>
-    /// Gets or sets the timeout for this task.Default is null (no timeout)
+    /// Gets the timeout for this task. Default is null (no timeout)
     /// </summary>
-    public TimeSpan? Timeout { get; set; }
+    public TimeSpan? Timeout { get; }
+
+    /// <summary>
+    /// Gets the name of the queue where this task should be executed.
+    /// Returns null to use the default queue or automatic routing for recurring tasks.
+    /// </summary>
+    public string? QueueName { get; }
 
     /// <summary>
     /// Gets or sets whether this task is CPU-bound. Default is false.